Complete exercises 1.14 and 1.15 and add folders
Exercises 1.14 and 1.15 are now complete. Orders of growth are still
strange to me so my answers may be very wrong.

I added a folder for the problem sets and book chapters. I also
discovered how to convert a latex file to pdf (`pdflatex FILE`, but
first I had to install the `texlive-latex-base` package (Debian)).